drainer: factor job & node watchers out of drainer.go
